Apply a better understanding to reviews-data
A user present within `latestReviews` is a indicator that we are not waiting on that user. But the state within `latestReviews` does not always reflect the last opinionated review of that user.

Hence we use the combination of `latestOpinionatedReviews` and `latestReviews`. `latestReviews` is used to figure out wether or not we are waiting for that user, whereas `latestOpinionatedReviews` is the review itself.

If a user is missing from `latestReviews` it means that although they are present in `latestOptionatedReviews`, we should treat them as a missing review. This is probably due to a "re-request review"